# services/platform-api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LesionCreate": {
        "additionalProperties": false,
        "properties": {
          "firstObservationId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Firstobservationid",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "required": [
          "firstObservationId"
        ],
        "title": "LesionCreate",
        "type": "object"
      },
      "LesionResponse": {
        "additionalProperties": false,
        "properties": {
          "anatomicalSite": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnatomicalSite"
              },
              {
                "type": "null"
              }
            ]
          },
          "confirmedObservationIds": {
            "items": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            },
            "title": "Confirmedobservationids",
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "lesionId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Lesionid",
            "type": "string"
          },
          "matchDecisionIds": {
            "items": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            },
            "title": "Matchdecisionids",
            "type": "array"
          },
          "region": {
            "$ref": "#/components/schemas/MouthRegion"
          },
          "status": {
            "$ref": "#/components/schemas/LesionStatus"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "lesionId",
          "region",
          "anatomicalSite",
          "label",
          "status",
          "confirmedObservationIds",
          "matchDecisionIds",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "title": "LesionResponse",
        "type": "object"
      },
      "LesionStatus": {
        "enum": [
          "tracking",
          "archived"
        ],
        "title": "LesionStatus",
        "type": "string"
      },
      "MatchDecisionCreate": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/MatchDecisionValue"
          },
          "rationale": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          }
        },
        "required": [
          "decision"
        ],
        "title": "MatchDecisionCreate",
        "type": "object"
      },
      "MatchDecisionResponse": {
        "additionalProperties": false,
        "properties": {
          "actorId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Actorid",
            "type": "string"
          },
          "decidedAt": {
            "format": "date-time",
            "title": "Decidedat",
            "type": "string"
          },
          "decidedBy": {
            "const": "patient",
            "default": "patient",
            "title": "Decidedby",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/MatchDecisionValue"
          },
          "decisionId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Decisionid",
            "type": "string"
          },
          "lesionId": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lesionid"
          },
          "proposalId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Proposalid",
            "type": "string"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          }
        },
        "required": [
          "decisionId",
          "proposalId",
          "decision",
          "actorId",
          "rationale",
          "decidedAt",
          "lesionId"
        ],
        "title": "MatchDecisionResponse",
        "type": "object"
      },
...
      "MatchProposalCreate": {
        "additionalProperties": false,
        "properties": {
          "candidateLesionId": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidatelesionid"
          },
          "candidatePriorObservationId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Candidatepriorobservationid",
            "type": "string"
          },
          "currentObservationId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Currentobservationid",
            "type": "string"
          },
          "expiresAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "modelVersions": {
            "additionalProperties": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            },
            "maxProperties": 64,
            "title": "Modelversions",
            "type": "object"
          },
          "proposalOrigin": {
            "$ref": "#/components/schemas/MatchProposalOrigin",
            "default": "automatic_model"
          },
          "rank": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "score": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "required": [
          "currentObservationId",
          "candidatePriorObservationId"
        ],
        "title": "MatchProposalCreate",
        "type": "object"
      },
...
      "MatchProposalResponse": {
        "additionalProperties": false,
        "properties": {
          "automaticallyConfirmed": {
            "const": false,
            "default": false,
            "title": "Automaticallyconfirmed",
            "type": "boolean"
          },
          "candidateLesionId": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidatelesionid"
          },
          "candidatePriorObservationId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Candidatepriorobservationid",
            "type": "string"
          },
          "currentObservationId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Currentobservationid",
            "type": "string"
          },
          "expiresAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "modelVersions": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Modelversions",
            "type": "object"
          },
          "proposalId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Proposalid",
            "type": "string"
          },
          "proposalOrigin": {
            "$ref": "#/components/schemas/MatchProposalOrigin"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "state": {
            "const": "proposed",
            "default": "proposed",
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "proposalId",
          "currentObservationId",
          "candidatePriorObservationId",
          "candidateLesionId",
          "proposalOrigin",
          "score",
          "rank",
          "modelVersions",
          "generatedAt",
          "expiresAt"
        ],
        "title": "MatchProposalResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Stoma3D Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/lesions": {
      "post": {
        "operationId": "create_lesion_v2_lesions_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LesionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LesionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Lesion",
        "tags": [
          "tracking"
        ]
      }
    },
    "/v2/lesions/{lesion_id}": {
      "get": {
        "operationId": "get_lesion_v2_lesions__lesion_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lesion_id",
            "required": true,
            "schema": {
              "title": "Lesion Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LesionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lesion",
        "tags": [
          "tracking"
        ]
      }
    },
    "/v2/match-decisions/{decision_id}": {
      "get": {
        "operationId": "get_match_decision_v2_match_decisions__decision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "title": "Decision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Match Decision",
        "tags": [
          "tracking"
        ]
      }
    },
    "/v2/match-proposals": {
      "post": {
        "operationId": "create_match_proposal_v2_match_proposals_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchProposalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Match Proposal",
        "tags": [
          "tracking"
        ]
      }
    },
    "/v2/match-proposals/{proposal_id}": {
      "get": {
        "operationId": "get_match_proposal_v2_match_proposals__proposal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Match Proposal",
        "tags": [
          "tracking"
        ]
      }
    },
    "/v2/match-proposals/{proposal_id}/decisions": {
      "post": {
        "operationId": "decide_match_proposal_v2_match_proposals__proposal_id__decisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchDecisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Match Proposal",
        "tags": [
          "tracking"
        ]
      }
    },
...
  }
}